requirements.txt
llama-index-core
llama-index-llms-openai
llama-index-llms-openai-like
llama-index-embeddings-huggingface
llama-index-embeddings-openai
pytest
pytest-mock
llama-index-vector-stores-chroma
chromadb
openai>=1.0.0
python-dotenv
pydantic>=2.0
pydantic-settings>=2.0
